# go/mysql: improve GTID encoding for OK packet