This isn't a super common path, but if we are tracking GTIDs, this will happen for every single query.

This enhancement effectively calculates the full size of the GTID data to be encoded up front to do a single allocation rather than naively appending to a slice.

I also added documentation and extracted the GTID encoding to it's own function with unit tests.

I also cleaned up how it's utilized to avoid some unnecessary byte -> string allocations within the packet writer.

```
$ benchstat {old,new}.txt
goos: darwin
goarch: arm64
pkg: vitess.io/vitess/go/mysql
               │   old.txt    │               new.txt               │
               │    sec/op    │   sec/op     vs base                │
EncGtidData-10   104.65n ± 1%   15.55n ± 0%  -85.14% (p=0.000 n=10)

               │   old.txt   │              new.txt               │
               │    B/op     │    B/op     vs base                │
EncGtidData-10   56.000 ± 0%   8.000 ± 0%  -85.71% (p=0.000 n=10)

               │  old.txt   │              new.txt               │
               │ allocs/op  │ allocs/op   vs base                │
EncGtidData-10   7.000 ± 0%   1.000 ± 0%  -85.71% (p=0.000 n=10)
```

## Checklist

-   [x] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [x] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required